pyproject.toml
[tool.poetry]
name = "pageloot-be-task"
version = "0.1.0"
description = ""
authors = ["k4rimdev <karimmirzaguliyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
django-jazzmin = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
